Move tools to a dedicated package
