+++ src/main/resources/org/jenkinsci/plugins/docker/commons/tools/DockerTool/config.jelly
