+++ src/test/resources/org/jenkinsci/plugins/docker/commons/util/SampleDockerBuilder/config.jelly
